backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0
